refactoring project. Move data_types/ and search/ to src/
