--- data_types/__init__.py
@@ -1,4 +1,0 @@
-from .post import Post
-from .values import Value, Sentence
-
-__all__ = ["Post", "Value", "Sentence"]
