+++ src/__init__.py
@@ -1,4 +1,2 @@
-from .likes import Like
-from .fvrt import Favorite
 
-__all__ = ["Like", "Favorite"]
+__all__ = []
