+++ src/data_types/__init__.py
@@ -1,0 +1,12 @@
+from .post import Post
+from .values import Value, Sentence
+from .likes import Like
+from .fvrt import Favorite
+
+__all__ = [
+    "Post",
+    "Value",
+    "Sentence",
+    "Like",
+    "Favorite"
+]
