+++ src/search/__init__.py
@@ -1,0 +1,7 @@
+from .searcher import Searcher
+from .elastic_searcher import ElasticSearcher
+
+__all__ = [
+    "Searcher",
+    "ElasticSearcher"
+]
